backend/requirements.txt
Django==1.3.1
Fabric==1.3.3
South==0.7.3
django-bcrypt==0.9.2
django-devserver==0.3.1
django-piston==0.2.3
django-treebeard==1.61
psycopg2==2.4.2
py-bcrypt==0.2
pycrypto==2.4.1
ssh==1.7.11
wsgiref==0.1.2
